requirements.txt
openai>=1.0.0
pyyaml>=6.0
tiktoken>=0.5.0
matplotlib>=3.7.0
streamlit>=1.28.0

